move functions that connect with ee back to common, then remove many unused functions from downloads.py, update the scripts, tests, and notebooks accordingly,
