+++ tests/test_common.py
@@ -17,6 +17,58 @@ import pandas as pd
 import pytest
 from unittest.mock import patch, call
 
+def test_authenticate_and_initialize_max_attempts():
+    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
+         patch('coastseg.common.ee.Initialize') as mock_initialize:
+        
+        # Mock an exception for all initialize attempts
+        mock_initialize.side_effect = Exception("Credentials file not found")
+        
+        with pytest.raises(Exception) as excinfo:
+            common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
+        
+        assert "Failed to initialize Google Earth Engine after 2 attempts" in str(excinfo.value)
+        assert mock_authenticate.call_count == 2
+        assert mock_initialize.call_count == 2
+
+
+def test_authenticate_and_initialize_success():
+    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
+         patch('coastseg.common.ee.Initialize') as mock_initialize:
+        
+        # Mock successful initialization
+        mock_initialize.return_value = None
+        
+        common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
+
+        mock_authenticate.assert_called_once() # this will call once becase ee.credentials is None
+        mock_initialize.assert_called_once()
+
+def test_authenticate_and_initialize_force_auth():
+    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
+         patch('coastseg.common.ee.Initialize') as mock_initialize:
+        
+        # Mock successful initialization
+        mock_initialize.return_value = None
+        
+        common.authenticate_and_initialize(print_mode=True, force=True, auth_args={}, kwargs={})
+
+        mock_authenticate.assert_called_once_with(force=True)
+        mock_initialize.assert_called_once()
+
+def test_authenticate_and_initialize_retry():
+    with patch('coastseg.common.ee.Authenticate') as mock_authenticate, \
+         patch('coastseg.common.ee.Initialize') as mock_initialize:
+        
+        # Mock an exception on first initialize, then success
+        mock_initialize.side_effect = [Exception("Credentials file not found"), None]
+
+        common.authenticate_and_initialize(print_mode=True, force=False, auth_args={}, kwargs={})
+
+        assert mock_authenticate.call_count == 2
+        assert mock_initialize.call_count == 2
+
+
 def test_empty_merged_timeseries_gdf():
     # Create an empty GeoDataFrame with the necessary structure
     empty_gdf = gpd.GeoDataFrame(columns=['x', 'y', 'shore_x', 'shore_y', 'cross_distance', 'dates'])
